requirements.txt
biopython
numpy
Pillow
pkg-resources
scipy
tqdm